voice-input-chatbot/requirements.txt
streamlit
groq
langchain
langchain-groq
python-dotenv
pyaudio
wave
speech_recognition
